requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
psycopg2==2.9.6
openai>=1.0.0
googlemaps==4.10.0
